fix(workspace): 周回をまたぐ fanout 子 Node の Workspace node id 衝突を解消する (#1702)
後方辺で fanout を再入する workflow や、同一 Worktree 内で同じ fanout を含む workflow を複数回実行すると、
fanout 子 Node の Workspace node id が衝突し、`corrupt_stored_state` エラーで Workspace tree 読み出しが
失敗していた。これにより UI が「Node unavailable」となり、ワークフローの観測や介入が不可能になっていた。

この変更により、fanout 子 Node の Workspace node id 導出に親 fanout インスタンスの NodeExecution id を
アンカーとして使用するよう改訂し、周回や複数実行をまたぐ一意性を保証する。
これにより、既存の fact log もデータ移行なしに正常に読み出せるようになり、UI からの観測と操作が可能になる。
