+++ src-tauri/src/adaptor/gateway/workspace_tree/repository.rs
@@ -284,56 +284,8 @@ pub(super) fn fold_query_error(reason: String) -> LocalEventQueryError {
 }
 
 #[cfg(test)]
-mod legacy_projection_tests {
-    use super::*;
-    use crate::adaptor::gateway::local_event_store::layout::StoreLayout;
-    use crate::adaptor::gateway::local_event_store::store::LocalEventStoreConfig;
-
-    #[tokio::test]
-    async fn legacy_agent_projection_row_is_ignored_by_canonical_session_and_workspace_queries() {
-        let root = tempfile::TempDir::new().unwrap();
-        let store =
-            LocalEventStore::open(LocalEventStoreConfig::production(root.path().to_path_buf()))
-                .unwrap();
-
-        let connection =
-            rusqlite::Connection::open(StoreLayout::new(root.path()).database_path()).unwrap();
-        connection
-            .execute(
-                "INSERT INTO logical_commits (
-                    commit_id, installation_id, operation_kind, idempotency_key, payload_hash,
-                    state, first_global_sequence, last_global_sequence, event_count,
-                    mutation_count, stream_heads_json, result_hash, committed_at_ms
-                 ) VALUES (?1, 'legacy-install', 'projection', 'legacy-key', ?2,
-                    'sealed', NULL, NULL, 0, 1, '[]', ?2, 0)",
-                rusqlite::params!["legacy-commit", [0_u8; 32].as_slice()],
-            )
-            .unwrap();
-        connection
-            .execute(
-                "INSERT INTO session_projection (
-                    session_id, projection, revision, commit_id, workspace_identity,
-                    public_list_kind, public_sort_key_bits, public_summary
-                 ) VALUES (?1, ?2, 0, ?3, ?4, 'active', 0, ?5)",
-                rusqlite::params![
-                    "legacy-session-1",
-                    r#"{"schema":"legacy_agent_session_projection_v0"}"#,
-                    "legacy-commit",
-                    "/repo",
-                    r#"{"schema":"legacy_session_public_summary_v0"}"#,
-                ],
-            )
-            .unwrap();
-        drop(connection);
-
-        let repository = SqliteWorkspaceTreeRepository::new(store);
-        let trees = repository.folded_workspace_trees("/repo").unwrap();
-        assert!(repository
-            .workspace_tree_from_folded("/repo", &trees)
-            .unwrap()
-            .is_none());
-    }
-}
+#[path = "legacy_projection_test.rs"]
+mod legacy_projection_tests;
 
 pub(super) fn codec_query_error(error: String) -> LocalEventQueryError {
     let correlation_id = uuid::Uuid::new_v4().to_string();
@@ -348,125 +300,5 @@ fn invariant_query_error(error: impl std::fmt::Display) -> LocalEventQueryError 
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::adaptor::gateway::agent_session::LocalAgentSessionRepository;
-    use crate::adaptor::gateway::local_event_store::store::LocalEventStoreConfig;
-    use crate::domain::agent_session::aggregates::AgentSession;
-    use crate::domain::agent_session::repository::AgentSessionRepository;
-    use crate::domain::provider_lifecycle::ProviderKind;
-
-    fn sqlite_failure(code: i32) -> rusqlite::Error {
-        rusqlite::Error::SqliteFailure(rusqlite::ffi::Error::new(code), None)
-    }
-
-    #[test]
-    fn sqlite_query_errors_preserve_store_and_record_failure_classification() {
-        for code in [rusqlite::ffi::SQLITE_CORRUPT, rusqlite::ffi::SQLITE_NOTADB] {
-            assert!(matches!(
-                sql_query_error(sqlite_failure(code)),
-                LocalEventQueryError::Corrupt { .. }
-            ));
-        }
-        for code in [rusqlite::ffi::SQLITE_BUSY, rusqlite::ffi::SQLITE_LOCKED] {
-            assert!(matches!(
-                sql_query_error(sqlite_failure(code)),
-                LocalEventQueryError::QueryBusy
-            ));
-        }
-        assert!(matches!(
-            sql_query_error(rusqlite::Error::FromSqlConversionFailure(
-                0,
-                rusqlite::types::Type::Text,
-                Box::new(std::io::Error::new(
-                    std::io::ErrorKind::InvalidData,
-                    "invalid value",
-                )),
-            )),
-            LocalEventQueryError::IncompatibleStoredEvent { .. }
-        ));
-        assert!(matches!(
-            sql_query_error(sqlite_failure(rusqlite::ffi::SQLITE_IOERR)),
-            LocalEventQueryError::StorageUnavailable { .. }
-        ));
-    }
-
-    #[tokio::test]
-    async fn public_session_root_id_loads_the_session_node_instead_of_the_internal_owner() {
-        let directory = tempfile::TempDir::new().unwrap();
-        let store = LocalEventStore::open(LocalEventStoreConfig::production(
-            directory.path().to_path_buf(),
-        ))
-        .unwrap();
-        let workspace = WorkspaceIdentity::new("/repo/.worktrees/feature");
-        let session = AgentSession::create(
-            "agent-session-1",
-            workspace.clone(),
-            workspace.as_str(),
-            ProviderKind::Codex,
-            None,
-        )
-        .unwrap();
-        LocalAgentSessionRepository::new(Arc::clone(&store))
-            .create(session, "create-request-1")
-            .await
-            .unwrap();
-        let repository = SqliteWorkspaceTreeRepository::new(store);
-
-        let node_id = repository
-            .node_id_for_session(&workspace, "agent-session-1")
-            .unwrap()
-            .expect("the standalone Session must have a public Node id");
-        let node = repository
-            .load_node(&workspace, &node_id)
-            .unwrap()
-            .expect("the public Node id must resolve");
-
-        assert_eq!(node_id, "agent-session-1");
-        assert_eq!(
-            node.kind,
-            crate::domain::workspace_tree::WorkspaceNodeKind::WorkflowSession
-        );
-        assert_eq!(node.node_execution_id.as_deref(), Some("agent-session-1"));
-        assert_eq!(node.session_id.as_deref(), Some("agent-session-1"));
-    }
-
-    #[tokio::test]
-    async fn a_session_owned_by_another_worktree_has_no_public_node_id() {
-        let directory = tempfile::TempDir::new().unwrap();
-        let store = LocalEventStore::open(LocalEventStoreConfig::production(
-            directory.path().to_path_buf(),
-        ))
-        .unwrap();
-        let owner = WorkspaceIdentity::new("/repo/.worktrees/feature");
-        let other = WorkspaceIdentity::new("/repo/.worktrees/other");
-        let session = AgentSession::create(
-            "agent-session-1",
-            owner.clone(),
-            owner.as_str(),
-            ProviderKind::Codex,
-            None,
-        )
-        .unwrap();
-        LocalAgentSessionRepository::new(Arc::clone(&store))
-            .create(session, "create-request-1")
-            .await
-            .unwrap();
-        let repository = SqliteWorkspaceTreeRepository::new(store);
-
-        assert!(
-            repository
-                .node_id_for_session(&other, "agent-session-1")
-                .unwrap()
-                .is_none(),
-            "another Worktree must not resolve a public Node id for this Session"
-        );
-        assert!(
-            repository
-                .node_id_for_session(&owner, "unknown-session")
-                .unwrap()
-                .is_none(),
-            "an unknown Session has no execution tree to publish"
-        );
-    }
-}
+#[path = "repository_test.rs"]
+mod repository_tests;
